refactor: redesign the edit event process
