+++ src/app/dashboard/e/[eventID]/booking/page.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Page = () => {
+  return (
+    <>
+      <h1 className="text-3xl font-medium">Booking Options</h1>
+    </>
+  );
+};
+
+export default Page;
